transfer shiftDetails and logs method to reusable service traits
